+++ apps/licences/src/utils/constants.ts
@@ -1,1 +1,3 @@
+export const CANCEL_CALLBACK_URL = "iossoapi://cancel";
+
 export const LICENCE_TOTAL_POINTS = 30;