+++ apps/vehicles/src/components/vehicle-details/VehicleSectionDetails.tsx
@@ -1,5 +1,9 @@
 import { ExtraMassaEnum, Veicolo } from "@/generated/bff-openapi";
-import { extraMassByCode, noviceByCode, vehicleByType } from "@/utils/strings";
+import {
+  extraMassByCode,
+  noviceByCode,
+  vehicleByType,
+} from "@/utils/constants";
 import { CardInfo, CardInfoItem, Icon, useDialog } from "@io-ipatente/ui";
 import { Stack, Typography } from "@mui/material";
 import { useTranslation } from "next-i18next";
@@ -110,10 +114,10 @@ const getExtraMassIcon = (massCode: ExtraMassaEnum) => {
     case ExtraMassaEnum.Enum.EXTRAM_MSG_002:
     case ExtraMassaEnum.Enum.EXTRAM_MSG_003:
     case ExtraMassaEnum.Enum.EXTRAM_MSG_004:
+    case ExtraMassaEnum.Enum.EXTRAM_MSG_006:
       return <Icon color="warning" fontSize="large" name="warning" />;
 
     case ExtraMassaEnum.Enum.EXTRAM_MSG_005:
-    case ExtraMassaEnum.Enum.EXTRAM_MSG_006:
       return <Icon color="error" fontSize="large" name="error" />;
   }
 };