+++ infra/repository/locals.tf
@@ -39,10 +39,11 @@ locals {
   }
 
   repository = {
-    name            = "io-ipatente"
-    description     = "iPatente services"
-    topics          = ["io", "ipatente"]
-    reviewers_teams = ["io-platform-green-unit", "engineering-team-cloud-eng"]
+    name               = "io-ipatente"
+    description        = "iPatente services"
+    topics             = ["io", "ipatente"]
+    reviewers_teams    = ["io-platform-green-unit", "engineering-team-cloud-eng"]
+    app_cd_policy_tags = ["io-ipatente-licences@*", "io-ipatente-payments@*", "io-ipatente-practices@*", "io-ipatente-vehicles@*"]
   }
 
   tags = {
@@ -54,3 +55,4 @@ locals {
     CostCenter     = "TS310 - PAGAMENTI & SERVIZI"
   }
 }
+