+++ packages/ui/lib/components/top-bar/TopBar.tsx
@@ -1,45 +1,36 @@
 import HelpOutlineRoundedIcon from "@mui/icons-material/HelpOutlineRounded";
-import { AppBar, IconButton, Link, Stack, Toolbar } from "@mui/material";
+import { AppBar, Box, IconButton, Stack, Toolbar } from "@mui/material";
 import { ButtonNaked } from "@pagopa/mui-italia";
 
-import { Logo, LogoType } from "../logo";
+import Logo from "../../assets/logo.png";
 
 interface AssistanceProps {
   label: string;
   onClick: () => void;
 }
 
-interface ProductProps {
-  logo: LogoType;
-  name: string;
-  url: string;
-}
-
 export interface TopBarProps {
   assistance?: AssistanceProps;
-  product: ProductProps;
 }
 
-export const TopBar = ({ assistance, product }: TopBarProps) => (
+export const TopBar = ({ assistance }: TopBarProps) => (
   <AppBar position="static" sx={{ bgcolor: "background.paper" }}>
     <Toolbar>
-      <Logo name={product.logo} />
-      <Stack
-        alignItems="center"
-        direction="row"
-        spacing={2}
-        sx={{ flexGrow: 1, ml: 1 }}
+      <Box
+        sx={{
+          alignItems: "center",
+          display: "flex",
+          flexGrow: 1,
+        }}
       >
-        <Link
-          color="primary"
-          href={product.url}
-          underline="none"
-          variant="caption"
-          width={100}
-        >
-          {product.name}
-        </Link>
-      </Stack>
+        <Box
+          component="img"
+          src={Logo}
+          sx={{
+            height: 38,
+          }}
+        />
+      </Box>
 
       {assistance && (
         <Stack alignItems="center" direction="row" spacing={2}>