+++ packages/core/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @io-ipatente/core
 
+## 0.6.1
+
+### Patch Changes
+
+- 0f765b0: Fix middleware
+- aec656e: Update Next to v14.2.20
+
+## 0.6.0
+
+### Minor Changes
+
+- f319d92: Add consent screen
+
+## 0.5.16
+
+### Patch Changes
+
+- 051fc4b: Minor fixes
+
 ## 0.5.15
 
 ### Patch Changes